Split CpDraw into 2 classes
